feat(app): expand GPU model and schema for richer hardware specs
Updates SQLModel + Pydantic shape for DiscreteGPU to expose the
additional fields populated by the rebuilt GPU dataset.

Refs #1

+++ app/models/gpu.py
@@ -28,7 +28,7 @@ class DiscreteGPU(SQLModel, table=True):
     tensor_cores: int | None = None
 
     # Memory
-    memory_gb: int
+    memory_gb: float
     memory_type: str
     memory_bus_bit: int
     memory_bandwidth_gbps: float | None = None
